Update BZAN-530-Hanes-Case-Study.Rproj
Created subsets and analyses to get the top 5 DMAs for each category in terms of most variance. Also, starting to do linear regression on factors for those top 5 DMAs with the best sales performance to look for correlation factors (lines 21-57). This is still just me brainstorming ways to figure all this out, so if you all have any suggestions, or see that I may be doing something wrong just let me know :)